Proof of concept for handling the Download fatal error
Note: I had to add a setTimeout on line 53 due to a race condition with the Zip API Call. We should not redirect to Step 3 until we get a valid Zip URL and verified that it won't produce any fatal errors. Once we've verified the URL we can point to step 3 and do the redirect.